Mapped chat help and status commands
Took 7 minutes
